core: don't use -1 as no-sensor indicator
In 82f967ddb37022a69be2de3f38ac445093c0d6df sanitizing of sensor-ids
was introduced to prevent out-of-bound accesses. The sensor was
overwritten as -1 (NO_SENSOR).

However, this led to corrupted XML files, since the code sees this
as an "old format" indicator.

Other parts of the code simply take the pressure and sensor reading
from the previous sample. Do this here as well.

Signed-off-by: Berthold Stoeger <bstoeger@mail.tuwien.ac.at>